requirements.txt
# Core AI and LangChain dependencies
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# Google Generative AI
google-generativeai>=0.3.0

# Vector search (lightweight version)
faiss-cpu>=1.7.0
# Using built-in embeddings from Google API instead of sentence-transformers

# Data processing and utilities
numpy>=1.21.0

# Web framework (for optional web interface)
flask>=2.3.0

# Date and time handling
python-dateutil>=2.8.0

# Environment management
python-dotenv>=0.19.0